+++ samplecode/SampleComplexClip.cpp
@@ -96,7 +96,7 @@ protected:
 
         for (int invA = 0; invA < 2; ++invA) {
             for (size_t op = 0; op < SK_ARRAY_COUNT(gOps); ++op) {
-                int idx = invA * SK_ARRAY_COUNT(gOps) + op;
+                size_t idx = invA * SK_ARRAY_COUNT(gOps) + op;
                 if (!(idx % 3)) {
                     canvas->restore();
                     canvas->translate(0, SkIntToScalar(250));
